changed yieldAmount form int to double. added button to submit final amount and complete project

+++ lib/models/plot_model.dart
@@ -21,7 +21,7 @@ class PlotModel {
   double predictedRevenue;
   double actualRevenue;
   String recommendedCrop;
-  int? yieldAmount;
+  double? yieldAmount;
 
   PlotModel({
     required this.plotId,
